+++ Assets/Scripts/Interfaces/IEngine.cs
@@ -3,6 +3,6 @@ interface IEngine
     public void PlayManual(int[] move);
     public int[] PlayAuto(int recursionDepth);
     public void SetState(byte[] newState);
-    public int Evaluate();
+    public float Evaluate();
     public bool CheckWin();
 }
